+++ TonieToolbox/__init__.py
@@ -2,4 +2,4 @@
 TonieToolbox - Convert audio files to Tonie box compatible format
 """
 
-__version__ = '0.6.0a1'
+__version__ = '0.6.0a2'